Split QuickEdits into their own files. Start working on TM Quick Edit again.
TM Quick Edit requires a data split and 6 known ASM changes. It's gonna be big, so it deserves its own file. Go ahead and give the other QuickEdits their own files too.
